Fix some abstract class problems
