+++ tests/Metadata/Driver/FileDriverTestCase.php
@@ -5,8 +5,6 @@ namespace Vich\UploaderBundle\Tests\Metadata\Driver;
 use Metadata\Driver\DriverInterface;
 use Metadata\Driver\FileLocatorInterface;
 use PHPUnit\Framework\TestCase;
-use Vich\TestBundle\Entity\Article;
-use Vich\TestBundle\Entity\Product;
 use Vich\UploaderBundle\Metadata\ClassMetadata;
 
 abstract class FileDriverTestCase extends TestCase
@@ -26,53 +24,6 @@ abstract class FileDriverTestCase extends TestCase
         self::assertEquals($expectedMetadata, $metadata->fields);
     }
 
-    public static function classesProvider(): array
-    {
-        $metadatas = [];
-        $metadatas[] = [
-            Product::class,
-            __DIR__ . '/../../Fixtures/TestBundle/config/vich_uploader/Entity.Product.' . self::getExtension(),
-            [
-                'image' => [
-                    'mapping' => 'product_image',
-                    'propertyName' => 'image',
-                    'fileNameProperty' => 'imageName',
-                    'size' => 'imageSize',
-                    'mimeType' => 'imageMimeType',
-                    'originalName' => 'imageOriginalName',
-                    'dimensions' => null,
-                ],
-            ],
-        ];
-
-        $metadatas[] = [
-            Article::class,
-            __DIR__ . '/../../Fixtures/TestBundle/config/vich_uploader/Entity.Article.' . self::getExtension(),
-            [
-                'attachment' => [
-                    'mapping' => 'dummy_file',
-                    'propertyName' => 'attachment',
-                    'fileNameProperty' => 'attachmentName',
-                    'size' => null,
-                    'mimeType' => null,
-                    'originalName' => null,
-                    'dimensions' => null,
-                ],
-                'image' => [
-                    'mapping' => 'dummy_image',
-                    'propertyName' => 'image',
-                    'fileNameProperty' => 'imageName',
-                    'size' => 'imageSize',
-                    'mimeType' => 'imageMimeType',
-                    'originalName' => 'imageOriginalName',
-                    'dimensions' => null,
-                ],
-            ],
-        ];
-
-        return $metadatas;
-    }
-
     protected function getFileLocatorMock(\ReflectionClass $class, ?string $foundFile = null): FileLocatorInterface
     {
         $fileLocator = $this->createMock(FileLocatorInterface::class);
@@ -84,6 +35,8 @@ abstract class FileDriverTestCase extends TestCase
 
         return $fileLocator;
     }
+
+    abstract static function classesProvider(): array;
 
     abstract static protected function getExtension(): string;
