+++ tests/Metadata/Driver/YamlDriverTest.php
@@ -4,6 +4,8 @@ namespace Vich\UploaderBundle\Tests\Metadata\Driver;
 
 use Metadata\Driver\DriverInterface;
 use Metadata\Driver\FileLocatorInterface;
+use Vich\TestBundle\Entity\Article;
+use Vich\TestBundle\Entity\Product;
 use Vich\UploaderBundle\Metadata\Driver\AbstractYamlDriver;
 use Vich\UploaderBundle\Metadata\Driver\YamlDriver;
 
@@ -12,6 +14,53 @@ use Vich\UploaderBundle\Metadata\Driver\YamlDriver;
  */
 final class YamlDriverTest extends FileDriverTestCase
 {
+    public static function classesProvider(): array
+    {
+        $metadatas = [];
+        $metadatas[] = [
+            Product::class,
+            __DIR__ . '/../../Fixtures/TestBundle/config/vich_uploader/Entity.Product.' . self::getExtension(),
+            [
+                'image' => [
+                    'mapping' => 'product_image',
+                    'propertyName' => 'image',
+                    'fileNameProperty' => 'imageName',
+                    'size' => 'imageSize',
+                    'mimeType' => 'imageMimeType',
+                    'originalName' => 'imageOriginalName',
+                    'dimensions' => null,
+                ],
+            ],
+        ];
+
+        $metadatas[] = [
+            Article::class,
+            __DIR__ . '/../../Fixtures/TestBundle/config/vich_uploader/Entity.Article.' . self::getExtension(),
+            [
+                'attachment' => [
+                    'mapping' => 'dummy_file',
+                    'propertyName' => 'attachment',
+                    'fileNameProperty' => 'attachmentName',
+                    'size' => null,
+                    'mimeType' => null,
+                    'originalName' => null,
+                    'dimensions' => null,
+                ],
+                'image' => [
+                    'mapping' => 'dummy_image',
+                    'propertyName' => 'image',
+                    'fileNameProperty' => 'imageName',
+                    'size' => 'imageSize',
+                    'mimeType' => 'imageMimeType',
+                    'originalName' => 'imageOriginalName',
+                    'dimensions' => null,
+                ],
+            ],
+        ];
+
+        return $metadatas;
+    }
+
     public function testInconsistentYamlFile(): void
     {
         $this->expectException(\RuntimeException::class);
